# identity-ms/identity-ms-api/swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "version" : "1.2.0-SNAPSHOT",
    "title" : "WSO2UserManagerAPI"
  },
  "basePath" : "/",
...
  "schemes" : [ "http", "https" ],
  "paths" : {
    "/admin/applications" : {
      "put" : {
        "tags" : [ "Application API" ],
        "summary" : "Update Application",
        "description" : "Returns HTTP 201 if the Application is successfully updated",
        "operationId" : "updateApplication",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "query",
          "required" : true,
          "type" : "string"
        }, {
          "in" : "body",
          "name" : "body",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/RoleUpdateDTO"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Valid Application"
          },
          "404" : {
            "description" : "Application is not found"
          }
        },
        "security" : [ {
          "X-Auth-Token" : [ ]
        } ]
      }
    },
...
    "/admin/resources" : {
      "patch" : {
        "tags" : [ "Authorization API" ],
        "summary" : "This function clears all the authorizations for the given resource",
        "description" : "",
        "operationId" : "clearResourceAuthorizations",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "resourceid",
          "in" : "query",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : ""
          }
        },
        "security" : [ {
          "X-Auth-Token" : [ ]
        } ]
      }
    },
    "/admin/roles" : {
      "post" : {
        "tags" : [ "Role API" ],
        "summary" : "Create new Role",
        "description" : "",
        "operationId" : "createRole",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/Role"
          }
        } ],
        "responses" : {
          "201" : {
            "description" : "Valid Role is created"
          },
          "404" : {
            "description" : "Role is not found"
          }
        },
        "security" : [ {
          "X-Auth-Token" : [ ]
        } ]
      },
      "delete" : {
        "tags" : [ "Role API" ],
        "summary" : "Delete a given role name",
        "description" : "",
        "operationId" : "deleteRole",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "rolename",
          "in" : "query",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "204" : {
            "description" : "Role is deleted"
          },
          "404" : {
            "description" : "Valid Role is not found"
          }
        },
        "security" : [ {
          "X-Auth-Token" : [ ]
        } ]
      },
      "patch" : {
        "tags" : [ "Authorization API" ],
        "summary" : "This function clears all authorizations of the role",
        "description" : "",
        "operationId" : "clearAllRoleAuthorization",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "rolename",
          "in" : "query",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : ""
          }
        },
        "security" : [ {
          "X-Auth-Token" : [ ]
        } ]
      }
    },
...
    "/admin/roles/resources" : {
      "put" : {
        "tags" : [ "Authorization API" ],
        "summary" : "This function authorizes the given role to perform the specified action on the given resource",
        "description" : "",
        "operationId" : "authorizeRole",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "rolename",
          "in" : "query",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "resourceid",
          "in" : "query",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : ""
          }
        },
        "security" : [ {
          "X-Auth-Token" : [ ]
        } ]
      },
      "delete" : {
        "tags" : [ "Authorization API" ],
        "summary" : "This function removes the authorization of the role to perform the given action on the specified resource",
        "description" : "",
        "operationId" : "denyRole",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "rolename",
          "in" : "query",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "resourceid",
          "in" : "query",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : ""
          }
        },
        "security" : [ {
          "X-Auth-Token" : [ ]
        } ]
      },
      "patch" : {
        "tags" : [ "Authorization API" ],
        "summary" : "This function clear the authorization of the specified role to perform the given action on the resource",
        "description" : "",
        "operationId" : "clearRoleAuthorization",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "rolename",
          "in" : "query",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "resourceid",
          "in" : "query",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : ""
          }
        },
        "security" : [ {
          "X-Auth-Token" : [ ]
        } ]
      }
    },
...
    "/applications" : {
      "get" : {
        "tags" : [ "Application API" ],
        "summary" : "Return applications according to the filter",
        "description" : "Returns HTTP 404 if the Applications are not found",
        "operationId" : "getApplications",
        "produces" : [ "application/json" ],
        "responses" : {
          "200" : {
            "description" : "Valid Applications are found",
            "schema" : {
              "type" : "array",
              "items" : {
                "type" : "string"
              }
            }
          },
          "404" : {
            "description" : "Valid Applications are not found"
          }
        },
        "security" : [ {
          "X-Auth-Token" : [ ]
        } ]
      }
    },
    "/applications/users" : {
      "get" : {
        "tags" : [ "Application API" ],
        "summary" : "Return Users belongs to some Application",
        "description" : "Returns HTTP 404 if the Users are not found",
        "operationId" : "getUsersOfApplications",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "query",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "Valid Users are found",
            "schema" : {
              "type" : "array",
              "items" : {
                "type" : "string"
              }
            }
          },
          "404" : {
            "description" : "Valid Users are not found"
          }
        },
        "security" : [ {
          "X-Auth-Token" : [ ]
        } ]
      }
    },
    "/resources" : {
      "get" : {
        "tags" : [ "Authorization API" ],
        "summary" : "This function retrieves the list of authorized roles to perform the given action on the specified resource",
        "description" : "",
        "operationId" : "getAllowedRolesForResource",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "resourceid",
          "in" : "query",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "Valid Roles are found",
            "schema" : {
              "type" : "array",
              "items" : {
                "type" : "string"
              }
            }
          }
        },
        "security" : [ {
          "X-Auth-Token" : [ ]
        } ]
      }
    },
...
    "/roles/resources" : {
      "get" : {
        "tags" : [ "Authorization API" ],
        "summary" : "This function checks whether the given role is authorized to perform the action on the specified resource",
        "description" : "",
        "operationId" : "isRoleAuthorized",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "rolename",
          "in" : "query",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "resourceid",
          "in" : "query",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : ""
          }
        },
        "security" : [ {
          "X-Auth-Token" : [ ]
        } ]
      }
    },
...
  },
  "securityDefinitions" : {
    "X-Auth-Token" : {
      "type" : "apiKey",
      "name" : "X-Auth-Token",
      "in" : "header"
    }
  },
  "definitions" : {
...
    "Role" : {
      "type" : "object",
      "required" : [ "id" ],
      "properties" : {
        "id" : {
          "type" : "string",
          "readOnly" : true
        },
        "newUsers" : {
          "type" : "array",
          "readOnly" : true,
          "items" : {
            "type" : "string"
          }
        },
        "permissions" : {
          "type" : "array",
          "readOnly" : true,
          "items" : {
            "type" : "string"
          }
        }
      }
    },
...
    "RoleUpdateDTO" : {
      "type" : "object",
      "properties" : {
        "newUsers" : {
          "type" : "array",
          "readOnly" : true,
          "items" : {
            "type" : "string"
          }
        },
        "deletedUsers" : {
          "type" : "array",
          "readOnly" : true,
          "items" : {
            "type" : "string"
          }
        }
      }
    },
...
  }
}